Fixed two bugs in the job manager.  The first one was because The "Serial" option was unable to deal with "Choices" properly.  The second was to convert vertically concatenated strings into cell arrays when reading in a job file.
Another change is to do with spatial normalisation.  This should really have been tested more, and is still work in progress. The changes were an attempt to invent a better heuristic to figure out a reasonable sampling distance. This started because the existing heuristic was based on how much additional smoothing was needed for the source or template images - which failed when no additional smoothing was applied.  It still fails, but some solution will be found.

SVN r218