Fixed the problems with processing 4D data that were pointed out by Roland Rutschmann.
The reason for the discrepancy is that SPM currently uses spm_write_vol.  Data
that is represented by integers (eg int16) has an associated scalefactor, and
spm_write_vol currently attempts to determine the appropriate scalefactor so
that the maximum range of integer values is used.  This works fine for 3D,
but for 4D, it can only rescale according to the intensities found in the
first 3D volume written.  This leads to truncation of the intensities if the
maximum intensity in subsequent volumes exceeds that of the first volume.

I've just created a fix for this, which is currently being tested.  Resliced
or spatially normalised data are now written with the same intensity scaling
as the original data.  Note that a possible side effect of this is that if
the dynamic range of the images is not very high, then smoothing is likely to
cause a loss of precision if the output is written as one of the integer
types. There should have been some incompatible scalefactor warnings
appearing as a result of the problem.

SVN r976