Fix reported by Philippe Pouliot: We would like to report an error in spm_hdm_ui. In line 212, legend('s','f','q','v',0); should be instead: legend('s','f','v','q',0);
the labels for deoxyhaemoglobin and volume are
inverted, as is clear from the specifications of
x(3) = volume and x(4) = deoxy in spm_fx_hdm.

This inversion is clear physiologically when looking at output
generated by spm_hdm_ui, as a standard positive BOLD
corresponds to an increase in flow and volume and a
decrease in deoxyhaemoglobin.

SVN r4048