New GUI utility for obtaining accurate total tissue volumes.
This uses a new brainmask.nii (actually an intracranial mask, which includes brain and CSF) based on the new (revision 5248) TPMs to exclude the eyes from the fluid tissue class. This mask is simply the largest connected component after thresholding at 0.5 the sum of the GM, WM and CSF TPMs; this can probably be improved with manual editing... I haven't yet updated the release notes tex file to describe this, as it might change further.

The routine uses modulated normalised segmentations (mwc*), creating them if necessary and deleting them afterwards if it created them, but also appends a variable to the seg8.mat file so that repeated runs don't need to recreate the mwc images unless the mask image has changed. The volumes can be used easily as dependencies in subsequent GLM specification.

SVN r5311