Add mean image to dependencies and outputs, if "Register to mean" is selected as estimation option.
Note that in the current implementation of the batch system this may
break jobs if "Register to mean" was selected, a dependency on the mean
image used later on and "Register to mean" deselected before running the
job. In this case, the batch system does not get notified that the mean
image output will not be computed.

SVN r5651