Enhancement: Changed 'return' to 'break' in spm_csd2gew.
Following a discussion in the MC meeting, it was noted that poor matrix conditioning in the Wilson-Burg algorithm can cause the maximum norm threshold to be exceeded. In this case, it is now more elegant to break the loop and return partial results, along with a warning message about the potential loss of numerical accuracy. This is preferable to simply exiting the loop without any results.