Bug Fix: Fixed the fieldmap alignment with the first volume of each session
For multi-session fMRI, the fieldmap is assumed to be aligned with the first
volume of the session. After realignment, the positioning of the fMRI is
adjusted, but that of the field map was not. Now modified the code so the
estimated movement of the first fMRI of the session is also applied to the
fieldmap. This is done in memory only and not on disk.
NB: Not tested yet