Bug fix: allocate pC.L with pE.L to keep fields order between structures.
In line 38 of spm_ssr_priors.m the lead field parameter L gets added to pE (if it was not already present). However, L does not get added to pC. In spm_csd_demo.m the latter happens in spm_L_priors.m which causes the overall order of L to be different between pE and pC. This misalignment in turn causes wrong computations later in the code.