MATLAB/Octave compatibility improvements
spm_check_version checks for multiple things
- Runtime environment (MATLAB, Octave, ...)
- Feature sets of runtime based on runtime version
- toolbox availability and toolbox versioning
This commit tries to disentangle these use cases in order to improve Octave compatibility.

Most of the feature and toolbox availability checks would need to be handled differently in Octave than in MATLAB. MATLAB versions and Octave versions do not sync, and toolboxes need a special packaging in Octave to  be recognised as such.

For now, a warning is issued whenever spm_check_version('matlab',...) is called on Octave, and MATLAB version checks always return true on Octave, with the exception of spm_check_version('matlab','online') which returns false on Octave.